refactor: Adjust to manage data and the current-id controlled by context with hook
