+++ src/App.jsx
@@ -1,11 +1,14 @@
 import { LinkProvider } from './contexts/LinkProvider';
+import { FolderProvider } from './contexts/FolderProvider';
 import MyRouter from './Router';
 
 const App = () => {
   return (
-    <LinkProvider>
-      <MyRouter />
-    </LinkProvider>
+    <FolderProvider>
+      <LinkProvider>
+        <MyRouter />
+      </LinkProvider>
+    </FolderProvider>
   );
 };
