+++ src/utils/useStoredData.js
@@ -6,7 +6,7 @@ export const useStoredData = (contextType) => {
   if (!context) {
     throw new Error('[PROVIDER ERROR]Provider 내부에서 사용하세요.');
   }
-  return context.storedData;
+  return { storedData: context.storedData, currentId: context.currentId };
 };
 
 export const useSetStoredData = (contextType) => {
@@ -16,5 +16,5 @@ export const useSetStoredData = (contextType) => {
     throw new Error('[PROVIDER ERROR]Provider 내부에서 사용하세요.');
   }
 
-  return context.setStoredData;
+  return { setStoredData: context.setStoredData, setCurrentId: context.setCurrentId };
 };
